Push image only after new image build
Without this change, a merge to the default branch
causes CI to attempt to push a test image.

The issue is that the test image isn't build on a
merge to the default branch.

The fix is to check that an image was built before
attempting to tag and push the image.

Also, separate the flush request from the upload
request in the publishing process.